requirements.txt
cryptography>=41.0.0
Pillow>=10.0.0
stegano>=0.11.0
PyQt6>=6.5.0
requests>=2.31.0
